feat: types create (#31)
- feat: types crate

- refactor(backend): use types crate instead of defining own types

- chore(types): remove main file
