+++ backend/src/meetup/mod.rs
@@ -1,3 +1,2 @@
 pub mod search;
 pub mod search_request;
-pub mod search_types;
