+++ types/src/meetup/mod.rs
@@ -1,0 +1,2 @@
+pub mod search_request;
+pub mod search_response;
